src/requirements.txt
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
SQLAlchemy==1.3.20
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
